+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Features Added
 
+- Implement success count network statsbeat
+    ([#25633](https://github.com/Azure/azure-sdk-for-python/pull/25633))
+
 ### Breaking Changes
 
 ### Bugs Fixed